Made dasboard dialog a reusable component

+++ frontend/src/components/dashboard/Archive.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect, useMemo } from "react";
-import api from "../api";
-import ArchiveDialog from "./ArchiveDialog";
-import { DataTable } from "../components/table/DataTable";
+import api from "../../api";
+import { DataTable } from "../table/DataTable";
 import {
   getCoreRowModel,
   getPaginationRowModel,
@@ -9,15 +8,16 @@ import {
 } from "@tanstack/react-table";
 import { archiveProps } from "./ArchiveColumns";
 import { getArchiveColumns } from "./ArchiveColumns";
-import Pagination from "../components/table/Pagination";
+import Pagination from "../table/Pagination";
 import { toast } from "sonner";
-import ConfirmDialog from "./ConfirmDialog";
+import ConfirmDialog from "../../pages/ConfirmDialog";
+import DashboardDialog from "./DashboardDialog";
 
 const Archive = () => {
   const [data, setData] = useState<archiveProps[]>([]);
   const [originalItem, setOriginalItem] = useState<archiveProps | null>(null);
   const [downloadName, setDownloadName] = useState<string | undefined>("");
-
+  const [confirmDelete, setConfirmDelete] = useState<boolean>(false); //state for opening the confirm delete dialog
   //sets the downloadName as the original file name from the backend; trigger changes whenever user types something new in the input element
   useEffect(() => {
     setDownloadName(originalItem?.title);
@@ -68,10 +68,26 @@ const Archive = () => {
     [setOriginalItem]
   );
 
-  const handleArchiveDelete = async (originalItem: archiveProps) => {
+  //gets the data from archive endpoint from the backend
+  useEffect(() => {
+    const getImage = async () => {
+      try {
+        const res = await api.get("/archive/");
+        console.log(res);
+        setData(res.data);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+    getImage();
+  }, [originalItem]);
+
+  const handleArchiveDelete = async () => {
     try {
+      if (!originalItem) return;
       await api.delete(`/archive/${originalItem.id}/`);
       setOriginalItem(null);
+      toast.success("Successfully deleted data");
     } catch (error) {
       console.log(error.message);
       console.log(error.response?.data);
@@ -79,14 +95,13 @@ const Archive = () => {
     }
   };
 
-  const handleArchiveEdit = async (
-    originalItem: archiveProps,
-    downloadName: string
-  ) => {
+  const handleArchiveEdit = async () => {
     try {
+      if (!originalItem) return;
       const image = await fetch(originalItem?.image); //get the url of the image
       const blob = await image.blob(); //convert the url to binary
 
+      if (!downloadName) return;
       //check if the filename ends with an extension ".png" or ".jpg"
       const fileNameEnd =
         downloadName.endsWith(".png") || downloadName.endsWith(".jpg");
@@ -114,20 +129,6 @@ const Archive = () => {
     }
   };
 
-  //gets the data from archive endpoint from the backend
-  useEffect(() => {
-    const getImage = async () => {
-      try {
-        const res = await api.get("/archive/");
-        console.log(res);
-        setData(res.data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    getImage();
-  }, [originalItem]);
-
   //initialization of the table that tanstack uses, along with its settings
   const table = useReactTable({
     data,
@@ -148,16 +149,27 @@ const Archive = () => {
         {/* table */}
         <DataTable table={table} />
         {/* dialog for viewing, updating, and deleting the archived images */}
-        <ArchiveDialog
+        <DashboardDialog
           image={originalItem?.image}
-          originalItem={originalItem}
-          setOriginalItem={setOriginalItem}
-          handleArchiveEdit={handleArchiveEdit}
-          handleArchiveDelete={handleArchiveDelete}
-          downloadName={downloadName}
-          setDownloadName={setDownloadName}
+          isOpen={originalItem}
+          setIsOpen={setOriginalItem}
+          fileName={downloadName}
+          setFileName={setDownloadName}
+          onConfirm={() => handleArchiveEdit()}
+          onCancel={() => setConfirmDelete(true)}
+          dialogTitle="Edit Image"
+          descriptionText="Update the image name below:"
+          buttonText="Delete?"
         />
 
+        <ConfirmDialog
+          confirmDelete={confirmDelete}
+          setConfirmDelete={setConfirmDelete}
+          title="Are you absolutely sure?"
+          descriptionText="This action cannot be undone. This will permanently delete your image from the server."
+          actionText="Delete"
+          onConfirm={() => handleArchiveDelete()}
+        />
         {/* pagination */}
         <div className="mt-2 mb-10">
           <Pagination table={table} />
